refactor: changing portfolio to projects and re-layout
