+++ .idea/workspace.xml
@@ -1,10 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ChangeListManager">
-    <list default="true" id="8610d892-4a60-4fda-a17a-a10cf1c7db0d" name="Default Changelist" comment="refactor: new flex skills bar">
+    <list default="true" id="8610d892-4a60-4fda-a17a-a10cf1c7db0d" name="Default Changelist" comment="fix: removing SKillsItem">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/skills/SkillsItem.js" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/skills/SkillsList.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/skills/SkillsList.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/Page.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/Page.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/portfolio/Portfolio.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/portfolio/Projects.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/portfolio/PortfolioItem.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/portfolio/ProjectsItem.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/portfolio/PortfolioList.js" beforeDir="false" afterPath="$PROJECT_DIR$/src/portfolio/ProjectsList.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/scss/App.scss" beforeDir="false" afterPath="$PROJECT_DIR$/src/scss/App.scss" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/scss/Portfolio.scss" beforeDir="false" afterPath="$PROJECT_DIR$/src/scss/Projects.scss" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/scss/Skills.scss" beforeDir="false" afterPath="$PROJECT_DIR$/src/scss/Skills.scss" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/scss/colors.scss" beforeDir="false" afterPath="$PROJECT_DIR$/src/scss/colors.scss" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -244,14 +250,7 @@
       <workItem from="1587389910766" duration="202000" />
       <workItem from="1615374200730" duration="107000" />
       <workItem from="1615374321866" duration="2951000" />
-      <workItem from="1615452077175" duration="18956000" />
-    </task>
-    <task id="LOCAL-00099" summary="updating data">
-      <created>1571683693573</created>
-      <option name="number" value="00099" />
-      <option name="presentableId" value="LOCAL-00099" />
-      <option name="project" value="LOCAL" />
-      <updated>1571683693573</updated>
+      <workItem from="1615452077175" duration="24213000" />
     </task>
     <task id="LOCAL-00100" summary="code : cleaning timeline data processing">
       <created>1571689633323</created>
@@ -589,7 +588,14 @@
       <option name="project" value="LOCAL" />
       <updated>1615476977491</updated>
     </task>
-    <option name="localTasksCounter" value="148" />
+    <task id="LOCAL-00148" summary="fix: removing SKillsItem">
+      <created>1615477154708</created>
+      <option name="number" value="00148" />
+      <option name="presentableId" value="LOCAL-00148" />
+      <option name="project" value="LOCAL" />
+      <updated>1615477154708</updated>
+    </task>
+    <option name="localTasksCounter" value="149" />
     <servers />
   </component>
   <component name="TodoView">
@@ -628,7 +634,6 @@
   </component>
   <component name="VcsManagerConfiguration">
     <option name="CHECK_CODE_CLEANUP_BEFORE_PROJECT_COMMIT" value="true" />
-    <MESSAGE value="fix : minor fixes on pdf and skills css" />
     <MESSAGE value="removing bootstrap" />
     <MESSAGE value="updating data" />
     <MESSAGE value="adding thumbnail" />
@@ -653,7 +658,8 @@
     <MESSAGE value="fix: clean header fonts and margins" />
     <MESSAGE value="fix: color of title and name" />
     <MESSAGE value="refactor: new flex skills bar" />
-    <option name="LAST_COMMIT_MESSAGE" value="refactor: new flex skills bar" />
+    <MESSAGE value="fix: removing SKillsItem" />
+    <option name="LAST_COMMIT_MESSAGE" value="fix: removing SKillsItem" />
     <option name="OPTIMIZE_IMPORTS_BEFORE_PROJECT_COMMIT" value="true" />
   </component>
   <component name="XDebuggerManager">
