+++ src/scss/App.scss
@@ -4,8 +4,14 @@ body {
   font-family: 'Helvetica Neue',Helvetica,Arial,sans-serif !important;
 }
 
-p {
+p, li, a {
   font-size: large;
+}
+
+
+time {
+  color: $theme-gray-700;
+  font-weight: bold;
 }
 
 a, .btn-link{
