--- src/scss/Portfolio.scss
@@ -1,86 +1,0 @@
-@import 'colors.scss';
-
-#portfolio {
-  p {
-    margin-bottom: 2px;
-  }
-
-  time {
-    color: $theme-gray-700;
-    font-size: 14px;
-    font-weight: bold;
-  }
-
-  .application,  .game, .web, .ip, .library, .simulator {
-    background-color: $theme-tag-red;
-  }
-
-  .computer.vision, .machine.learning, .side-channel  {
-    background-color: $theme-tag-purple;
-  }
-
-  .python, .rust, .typescript, .c\+\+, .c, .javascript {
-    background-color: $theme-tag-lemon;
-  }
-
-  .numpy, .scipy, .piston, .react, .npm, .fpga {
-    background-color: $theme-tag-orange;
-  }
-
-  .portfolio-item-show & {
-    flex-grow: 10;
-  }
-
-  .vhdl {
-    background-color: $theme-tag-blue;
-  }
-}
-
-.portfolio-tag {
-  margin-right: 12px;
-  margin-bottom: 12px;
-  font-size: 12px;
-  font-weight: bold;
-  letter-spacing: 1px;
-  text-transform: uppercase;
-  border-radius: 10px;
-  color: white;
-  padding: 4px;
-  background-color: $theme-gray-700;
-}
-
-.portfolio-container {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: center;
-}
-
-.portfolio-item {
-  margin: 24px 0 24px 24px;
-  padding: 16px;
-  background-color: $theme-gray-300;
-  border-radius: 10px;
-  flex-grow: 1;
-}
-
-@media only screen and (max-width: 768px) {
-  .portfolio-item {
-    flex-basis: 100%;
-  }
-}
-
-@media only screen and (min-width: 768px) {
-  .portfolio-item {
-    flex-basis: 40%;
-  }
-}
-
-
-.portfolio-bottom {
-  margin-top: 8px;
-}
-
-.portfolio-text {
-  font-style: italic;
-}
